backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pypdf==3.17.1
aiofiles==23.2.1
numpy==1.26.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.1
rank-bm25==0.2.2
pandas==2.1.4
openpyxl==3.1.2
